PMM-544 mostly structural fixes
