+++ pmm-app/src/pmm-inventory/ColumnRenderers/index.ts
@@ -1,0 +1,1 @@
+export * from './ColumnRenderers';
