+++ pmm-app/src/pmm-inventory/Inventory.types.ts
@@ -1,23 +1,38 @@
 export enum InventoryType {
+  amazonRdsMysql = 'amazon_rds_mysql',
+  container = 'container',
   externalExporter = 'externalExporter',
+  generic = 'generic',
+  mongodb = 'mongodb',
   mongodbExporter = 'mongodbExporter',
+  mysql = 'mysql',
   mysqldExporter = 'mysqldExporter',
-  postgresExporter = 'postgresExporter',
-  proxysqlExporter = 'proxysqlExporter',
   nodeExporter = 'nodeExporter',
   pmmAgent = 'pmm_agent',
+  postgresExporter = 'postgresExporter',
+  postgresql = 'postgresql',
+  proxysql = 'proxysql',
+  proxysqlExporter = 'proxysqlExporter',
   qanMongodb_profiler_agent = 'qan_mongodb_profiler_agent',
   qanMysql_perfschema_agent = 'qan_mysql_perfschema_agent',
   qanMysql_slowlog_agent = 'qan_mysql_slowlog_agent',
   qanPostgresql_pgstatements_agent = 'qan_postgresql_pgstatements_agent',
   rdsExporter = 'rdsExporter',
-  container = 'container',
-  generic = 'generic',
   remote = 'remote',
   remote_rds = 'remote_rds',
-  amazonRdsMysql = 'amazon_rds_mysql',
-  mongodb = 'mongodb',
-  mysql = 'mysql',
-  postgresql = 'postgresql',
-  proxysql = 'proxysql',
+}
+
+export type InventoryList = {
+  [key in InventoryType]: Node[];
+};
+
+export type ServicesList = InventoryList;
+
+export type NodesList = InventoryList;
+
+export type AgentsList = InventoryList;
+
+export interface CustomLabel {
+  key: string;
+  value: string;
 }
