+++ pmm-app/src/pmm-inventory/Tabs/index.ts
@@ -1,3 +1,3 @@
-export { Agents } from './agents/Agents';
-export { NodesTab } from './nodes/Nodes';
-export { Services } from './services/Services';
+export { Agents } from './Agents';
+export { NodesTab } from './Nodes';
+export { Services } from './Services';
