+++ pmm-app/src/pmm-inventory/panel.constants.tsx
@@ -1,7 +1,4 @@
-import React from 'react';
-
-export const getCustomLabels = customLabels =>
-  customLabels.map(label => <span key={label.key}>{`${label.key}: ${label.value}`}</span>);
+import { servicesDetailsRender, agentsDetailsRender, nodesDetailsRender } from './ColumnRenderers';
 
 export const MAIN_COLUMN = [
   'service_id',
@@ -12,51 +9,6 @@ export const MAIN_COLUMN = [
   'address',
   'port',
 ];
-
-const servicesDetailsRender = element => {
-  const labels = Object.keys(element).filter(label => !MAIN_COLUMN.includes(label));
-  return (
-    <div className="other-details-wrapper">
-      {labels.map((label, accessor) =>
-        element[label] ? <span key={accessor}>{`${label}: ${element[label]}`}</span> : null
-      )}
-      {element.custom_labels && getCustomLabels(element.custom_labels)}
-    </div>
-  );
-};
-
-export const agentsDetailsRender = element => {
-  const mainColumns = ['agent_id', 'type', 'isDeleted', 'service_ids', 'custom_labels'];
-  const labels = Object.keys(element).filter(label => !mainColumns.includes(label));
-  return (
-    <div className="other-details-wrapper">
-      {labels.map((label, key) =>
-        element[label] ? <span key={key}>{`${label}: ${element[label]}`}</span> : null
-      )}
-      {element.username ? <span>password: ******</span> : null}
-      {element.service_ids && element.service_ids.length ? (
-        /* TODO(lunaticusgreen): all map renders need a key! */
-        <>
-          service_ids: <span>{element.service_ids.map(serviceId => serviceId)}</span>
-        </>
-      ) : null}
-      {element.custom_labels && getCustomLabels(element.custom_labels)}
-    </div>
-  );
-};
-
-export const nodesDetailsRender = element => {
-  const mainColumns = ['node_id', 'node_name', 'address', 'custom_labels', 'type', 'isDeleted'];
-  const labels = Object.keys(element).filter(label => !mainColumns.includes(label));
-  return (
-    <div className="other-details-wrapper">
-      {labels.map((label, key) =>
-        element[label] ? <span key={key}>{`${label}: ${element[label]}`}</span> : null
-      )}
-      {element.custom_labels && getCustomLabels(element.custom_labels)}
-    </div>
-  );
-};
 
 export const SERVICES_COLUMNS = [
   {
