+++ pmm-app/src/react-plugins-deps/components/Table/index.ts
@@ -1,0 +1,2 @@
+export { Table } from './Table';
+export * from './Table.types';
